requirements.txt
PySide6>=6.0.0
Pillow>=8.0.0
pandas>=1.3.0
openpyxl>=3.0.0
pytest>=7.0.0
markdown>=3.3.0 